started wiring up the editor
currently wiring up the editor to stay in sync with state changes, especially when navigating between questions.

difficulty and approved state are wired up to navigation correctly. currently there's a bug in navigating between questions, where the EditableQuestion component does not sync up with the content of the question. address.