refactored Button functions
refactored the Button so that its SuperBigAndSpecial variant is a separate component. it should take the exact same set of props as the normal Button component.

created LoadingSpinner.jsx, which is a simple SVG animation that can be put into other components. it can take a prop optionalText that will display text underneath it as it spins.