changes to scripts.ts, questionModel.ts, Questions.ts
created new mongodb collection "gpt-questions--approved", which the front-end now sends to the server.

scripts.ts now parses the gpt completion with a fallback. if that fails, it provides dummy answers.

fixed bugs in the API endpoints related to async (did not properly use the 'await' keyword for some database requests).