changes to Requests.ts, created utils.ts
in Requests.ts,
created a dummy endpoint (/requests/test-difficulty-evaluator) to test whether the difficulty evaluator using ARI is working properly. currently is.

in utils.ts,
created three functions to help with evaluating the difficulty of text. these functions will be used in Requests.ts after GPT completions are done and before those questions are sent to the reader.