changes to utils.ts, index.jsx, Editor.jsx
in utils.ts,
moved around the weights for the reading difficulty.

in index.jsx,
removed <StrictMode> as it currently interferes with the random nature of GPT's generations (should not be a problem OUTSIDE of tests and generations!)

in Editor.jsx,
uncommented some code that prevented the editor from making the intended call to the right endpoint.